% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{add_cytoband_annotations}
\alias{add_cytoband_annotations}
\title{Add Cytoband Annotations to Targets}
\usage{
add_cytoband_annotations(targets, reference)
}
\arguments{
\item{targets}{Targets data.table}

\item{reference}{Reference object}
}
\value{
Targets with cytoband annotations
}
\description{
Add Cytoband Annotations to Targets
}
\keyword{internal}
