% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{add_exon_annotations}
\alias{add_exon_annotations}
\title{Add Exon Annotations to Targets}
\usage{
add_exon_annotations(targets, reference)
}
\arguments{
\item{targets}{Targets data.table}

\item{reference}{Reference object}
}
\value{
Targets with exon annotations
}
\description{
Add Exon Annotations to Targets
}
\keyword{internal}
