% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{add_gene_annotations}
\alias{add_gene_annotations}
\title{Add Gene Annotations to Targets}
\usage{
add_gene_annotations(targets, reference)
}
\arguments{
\item{targets}{Targets data.table}

\item{reference}{Reference object}
}
\value{
Targets with gene annotations
}
\description{
Add Gene Annotations to Targets
}
\keyword{internal}
