% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_genomic_positions}
\alias{add_genomic_positions}
\title{Add Genomic Positions to Targets and Segments}
\usage{
add_genomic_positions(targets, segments, chroms)
}
\arguments{
\item{targets}{data.table}

\item{segments}{data.table}

\item{chroms}{data.table with chromosome lengths}
}
\value{
List with modified targets and segments
}
\description{
Calculate genomic coordinates across chromosomes
}
\keyword{internal}
