% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somatic.R
\name{annotate_hotspots}
\alias{annotate_hotspots}
\title{Annotate Hotspots and Variant Effects}
\usage{
annotate_hotspots(somatic, genome)
}
\arguments{
\item{somatic}{A filtered, CSQ-annotated somatic data.table.}

\item{genome}{Genome build version ("hg19" or "hg38").}
}
\value{
The mutated data.table with `hotkey`, `is_hotspot`, and `effect` columns.
}
\description{
Maps somatic mutations to known hotspot tables and classifies their impact.
}
\keyword{internal}
