% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{annotate_segments}
\alias{annotate_segments}
\title{Annotate Segments with Cancer Genes and Exons}
\usage{
annotate_segments(
  segments,
  cancergenes,
  cancerexons,
  allgenes = NULL,
  cytobands = NULL
)
}
\arguments{
\item{segments}{Processed segments data.table.}

\item{cancergenes}{Data.table of cancer genes with coordinates.}

\item{cancerexons}{Data.table of cancer exons with coordinates.}

\item{allgenes}{Data.table of all protein-coding genes.}

\item{cytobands}{Data.table of cytoband coordinates.}
}
\value{
The annotated `segments` data.table.
}
\description{
Finds overlaps between derived genomic segments and known cancer genes/exons,
appending relevance labels to the segments.
}
\keyword{internal}
