% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{annotate_tiled_bins}
\alias{annotate_tiled_bins}
\title{Annotate Tiled Bins}
\usage{
annotate_tiled_bins(targets)
}
\arguments{
\item{targets}{Targets data.table}
}
\value{
Targets with tiled annotations
}
\description{
Annotate Tiled Bins
}
\keyword{internal}
