% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{apply_bin_median_correction}
\alias{apply_bin_median_correction}
\title{Apply Bin Median Correction}
\usage{
apply_bin_median_correction(targets, lr_col, ref_col)
}
\arguments{
\item{targets}{data.table}

\item{lr_col}{Column name with log ratios}

\item{ref_col}{Name for output reference median column}
}
\value{
Modified targets with reference-corrected values
}
\description{
Subtract bin-wise median (reference correction)
}
\keyword{internal}
