% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{apply_value_floor}
\alias{apply_value_floor}
\title{Apply Value Floor}
\usage{
apply_value_floor(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Vector with floor of 1 and NA replaced
}
\description{
Helper to replace values below 1 with 1 (for log safety)
}
\keyword{internal}
