% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{apply_x_chromosome_correction}
\alias{apply_x_chromosome_correction}
\title{Apply X Chromosome Correction}
\usage{
apply_x_chromosome_correction(targets, is_wgs)
}
\arguments{
\item{targets}{Targets data.table}

\item{is_wgs}{Logical indicating if WGS data}
}
\value{
Corrected targets
}
\description{
Apply X Chromosome Correction
}
\keyword{internal}
