% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{attach_ideogram}
\alias{attach_ideogram}
\title{Attach Ideogram using Inset}
\usage{
attach_ideogram(
  p,
  ideogram,
  strip_height_cm = 0.3,
  strip_margin_cm = 0.08,
  strip_title = TRUE
)
}
\arguments{
\item{p}{ggplot object}

\item{ideogram}{ggplot basic ideogram}

\item{strip_height_cm}{Height of ideogram}

\item{strip_margin_cm}{Margin gap between plot border and ideogram}

\item{strip_title}{Logical: whether to strip the axis.title.x label}
}
\value{
ggplot object with inset attached
}
\description{
Attach Ideogram using Inset
}
\keyword{internal}
