% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{blue}
\alias{blue}
\title{Blue colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
blue
}
\description{
Blue colour
}
\examples{
scales::show_col(blue)
}
\keyword{datasets}
