% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{build_ideogram_strip}
\alias{build_ideogram_strip}
\title{Build Ideogram Strip}
\usage{
build_ideogram_strip(chroms)
}
\arguments{
\item{chroms}{data.table with chromosome lengths}
}
\value{
ggplot object
}
\description{
Build Ideogram Strip
}
\keyword{internal}
