% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{calculate_gc_content}
\alias{calculate_gc_content}
\title{Calculate GC Content}
\usage{
calculate_gc_content(targets, ucsc_ranges, genome)
}
\value{
A data.table with GC-content values added for genomic bins.
}
\description{
Calculate GC Content
}
\keyword{internal}
