% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{calculate_logr}
\alias{calculate_logr}
\title{Calculate Log Ratio}
\usage{
calculate_logr(targets, count_col, output_col)
}
\arguments{
\item{targets}{data.table with count column}

\item{count_col}{Column name to convert}

\item{output_col}{Name for output column}
}
\value{
Modified targets with output_col added
}
\description{
Convert counts to log2 ratio with safety checks
}
\keyword{internal}
