% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{clamp_logr_values}
\alias{clamp_logr_values}
\title{Clamp Log Ratio Values}
\usage{
clamp_logr_values(targets, lr_col, min_val, max_val)
}
\arguments{
\item{targets}{data.table}

\item{lr_col}{Column to clamp}

\item{min_val}{Minimum value}

\item{max_val}{Maximum value}
}
\value{
Modified targets
}
\description{
Set min/max bounds on log ratios
}
\keyword{internal}
