% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{cleanup_gender_columns}
\alias{cleanup_gender_columns}
\title{Clean Temporary Gender Detection Columns}
\usage{
cleanup_gender_columns(targets)
}
\arguments{
\item{targets}{data.table}
}
\value{
Modified targets with temp columns removed
}
\description{
Clean Temporary Gender Detection Columns
}
\keyword{internal}
