% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{compute_gc_bias}
\alias{compute_gc_bias}
\title{Compute GC Bias Metric}
\usage{
compute_gc_bias(targets)
}
\arguments{
\item{targets}{Targets data.table}
}
\value{
Numeric GC bias value (log2 ratio) or NA
}
\description{
Compute GC Bias Metric
}
\keyword{internal}
