% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{compute_noise}
\alias{compute_noise}
\title{Compute Noise (Linear MAPD)}
\usage{
compute_noise(target_bins)
}
\arguments{
\item{target_bins}{Target bins data.table}
}
\value{
Numeric noise value or NA
}
\description{
Compute Noise (Linear MAPD)
}
\keyword{internal}
