% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{compute_snp_stats}
\alias{compute_snp_stats}
\title{Compute Het/Hom SNP Counts and Infer Sex}
\usage{
compute_snp_stats(snp_table, targets)
}
\arguments{
\item{snp_table}{SNP data.table (NULL if no SNP VCF)}

\item{targets}{Targets data.table (for counting X target bins)}
}
\value{
List with het_snps, hom_snps, sex
}
\description{
Compute Het/Hom SNP Counts and Infer Sex
}
\keyword{internal}
