% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{compute_waviness}
\alias{compute_waviness}
\title{Compute Waviness (1Mb Window Smoothed SD)}
\usage{
compute_waviness(target_bins)
}
\arguments{
\item{target_bins}{Target bins data.table}
}
\value{
Numeric waviness value or NA
}
\description{
Compute Waviness (1Mb Window Smoothed SD)
}
\keyword{internal}
