% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{correct_by_pca}
\alias{correct_by_pca}
\title{Apply PCA Correction to Values}
\usage{
correct_by_pca(data, train_indices = NULL)
}
\arguments{
\item{data}{data.table with 'lr' and PC columns}

\item{train_indices}{Logical or numeric indices for training subset}
}
\value{
Corrected log ratio vector
}
\description{
Robust linear model correction using PCA components
}
\keyword{internal}
