% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{create_background_bins}
\alias{create_background_bins}
\title{Create Background (Antitarget) Bins}
\usage{
create_background_bins(raw_targets, chromlength)
}
\value{
A GRanges object containing background bins.
}
\description{
Create Background (Antitarget) Bins
}
\keyword{internal}
