% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_snp_plots}
\alias{create_snp_plots}
\title{Create SNP-related Plots}
\usage{
create_snp_plots(
  targets,
  cancergenes_summary,
  chroms,
  chroms_order,
  somatic_plot,
  somatic_labels,
  yaxis_params,
  theme_params
)
}
\arguments{
\item{targets}{data.table with SNP data}

\item{cancergenes_summary}{data.table}

\item{chroms}{data.table}

\item{chroms_order}{data.table}

\item{somatic_plot}{data.table or NULL}

\item{somatic_labels}{data.table or NULL}

\item{yaxis_params}{List from prepare_yaxis_params()}

\item{theme_params}{List from setup_plot_theme()}
}
\value{
List of ggplot objects
}
\description{
Creates depth vs allele ratio, grid, nogrid, order, and position plots
}
\keyword{internal}
