% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{create_target_bins}
\alias{create_target_bins}
\title{Create Target Bins}
\usage{
create_target_bins(raw_targets)
}
\value{
A GRanges object containing target bins.
}
\description{
Create Target Bins
}
\keyword{internal}
