% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{create_target_template}
\alias{create_target_template}
\title{Create Target Template & Order Data}
\usage{
create_target_template(counts1, is_wgs, allcounts)
}
\value{
A list containing the target template and reordered count objects.
}
\description{
Create Target Template & Order Data
}
\keyword{internal}
