% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{create_wgs_bins}
\alias{create_wgs_bins}
\title{Create WGS Bins}
\usage{
create_wgs_bins(chromlength, bin_size)
}
\value{
A GRanges object containing WGS bins.
}
\description{
Create WGS Bins
}
\keyword{internal}
