% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_genome}
\alias{detect_genome}
\title{Detect Genome Build from BAM Header}
\usage{
detect_genome(bam_file)
}
\arguments{
\item{bam_file}{Path to the BAM file.}
}
\value{
A string "hg19", "hg38", or NULL if undetermined.
}
\description{
Inspects the BAM header to guess the genome build (hg19 or hg38).
}
\keyword{internal}
