% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{detect_pca_outliers}
\alias{detect_pca_outliers}
\title{Detect and Flag PCA Outliers}
\usage{
detect_pca_outliers(pca_data, sd_factor = 3, top_pcs = 100)
}
\arguments{
\item{pca_data}{data.table from perform_pca_on_bins}

\item{sd_factor}{Multiplier for standard deviation threshold}

\item{top_pcs}{Number of PCs to inspect (default 100)}
}
\value{
Vector of outlier bins (empty if none)
}
\description{
Identify bins with extreme PCA scores (outliers)
}
\keyword{internal}
