% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contamination.R
\name{estimate_contamination}
\alias{estimate_contamination}
\title{Estimate DNA Contamination}
\usage{
estimate_contamination(snp_table)
}
\arguments{
\item{snp_table}{A data.table containing SNP data, including 'allele_ratio'.}
}
\value{
A numeric value representing the estimated contamination fraction (0 to 1).
}
\description{
Estimates the level of DNA contamination in a sample based on SNP allele ratios.
}
