% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somatic.R
\name{extract_base_variants}
\alias{extract_base_variants}
\title{Extract Base Somatic Variants and Allele Metrics}
\usage{
extract_base_variants(vcf, tumor_idx)
}
\arguments{
\item{vcf}{A VCF object.}

\item{tumor_idx}{Integer indicating the column index of the tumor sample.}
}
\value{
A data.table containing unannotated somatic variants.
}
\description{
Extracts chromosome, position, reference, and alternate alleles from a VCF, 
and calculates Allele Frequency (AF), Allele Observation (AO), and Depth (DP).
}
\keyword{internal}
