% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{extract_snp_table}
\alias{extract_snp_table}
\title{Extract and Filter SNP Table from VCF}
\usage{
extract_snp_table(vcf)
}
\arguments{
\item{vcf}{A single-sample VCF object.}
}
\value{
A processed data.table containing SNP information.
}
\description{
Filters a single-sample VCF for biallelic SNPs and constructs a data.table 
containing allele depths, calculating allele ratios.
}
\keyword{internal}
