% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{extract_valid_bins}
\alias{extract_valid_bins}
\title{Filter Valid Bins}
\usage{
extract_valid_bins(matrix)
}
\arguments{
\item{matrix}{data.table with bin column}
}
\value{
Vector of valid bins
}
\description{
Identify bins that survived QC filtering
}
\keyword{internal}
