% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{filter_bins_by_coverage}
\alias{filter_bins_by_coverage}
\title{Filter Bins by Coverage Thresholds}
\usage{
filter_bins_by_coverage(targets)
}
\arguments{
\item{targets}{data.table with count column}
}
\value{
Filtered targets
}
\description{
Remove bins with extremely low or high coverage
}
\keyword{internal}
