% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{format_plot_legends}
\alias{format_plot_legends}
\title{Add Legend Formatting to Plots}
\usage{
format_plot_legends(plot_list)
}
\arguments{
\item{plot_list}{List of ggplot objects}
}
\value{
Modified list with consistent legend formatting
}
\description{
Add Legend Formatting to Plots
}
\keyword{internal}
