% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{generate_counts_from_bam}
\alias{generate_counts_from_bam}
\title{Generate Counts from BAM File}
\usage{
generate_counts_from_bam(bam_file, reference, output_dir)
}
\arguments{
\item{bam_file}{Path to BAM file}

\item{reference}{Reference object}

\item{output_dir}{Output directory}
}
\value{
Counts list
}
\description{
Generate Counts from BAM File
}
\keyword{internal}
