% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{generate_gene_annotation}
\alias{generate_gene_annotation}
\title{Generate Gene Annotation}
\usage{
generate_gene_annotation(genome = "hg19", mirror = NULL)
}
\arguments{
\item{genome}{Genome version ("hg19" or "hg38").}
}
\value{
A list containing `cancergenes`, `allgenes`, and `allexons`.
}
\description{
Fetches gene and exon information from Ensembl via biomaRt and integrates with
a bundled CSV file of clinically relevant cancer genes.
}
\keyword{internal}
