% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_annotation.R
\name{generate_gene_annotation_manual}
\alias{generate_gene_annotation_manual}
\title{Manual Gene Annotation Loader}
\usage{
generate_gene_annotation_manual(genome = "hg38")
}
\description{
Manual Gene Annotation Loader
}
