% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{get_bam_info}
\alias{get_bam_info}
\title{Extract Chromosome Lengths and Style from BAM}
\usage{
get_bam_info(bam_file)
}
\value{
A list containing chromosome lengths and chromosome naming style metadata.
}
\description{
Extract Chromosome Lengths and Style from BAM
}
\keyword{internal}
