% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{get_bins_by_type}
\alias{get_bins_by_type}
\title{Extract Bin Subset for PCA}
\usage{
get_bins_by_type(target_template, is_target_bool, valid_bins = NULL)
}
\arguments{
\item{target_template}{data.table with bin column}

\item{is_target_bool}{Logical: TRUE for targets, FALSE for background}

\item{valid_bins}{Bins that passed QC}
}
\value{
Vector of bins of specified type
}
\description{
Get bins of specific type (target or background)
}
\keyword{internal}
