% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{get_chrom_arms}
\alias{get_chrom_arms}
\title{Get Chromosome Arms}
\usage{
get_chrom_arms(genome = "hg19")
}
\value{
A data.table containing chromosome arm coordinates.
}
\description{
Get Chromosome Arms
}
\keyword{internal}
