% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_ensembl_mart}
\alias{get_ensembl_mart}
\title{Get Ensembl Mart Setup}
\usage{
get_ensembl_mart(genome, mirror = NULL)
}
\value{
A biomaRt Mart object.
}
\description{
Get Ensembl Mart Setup
}
\keyword{internal}
