% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_jumble_version}
\alias{get_jumble_version}
\title{Get Jumble Package Version}
\usage{
get_jumble_version()
}
\value{
A character string with the version number.
}
\description{
Robustly retrieves the package version, even when loaded via devtools.
}
\keyword{internal}
