% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{grey}
\alias{grey}
\title{grey colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
grey
}
\description{
grey colour
}
\examples{
scales::show_col(grey)
}
\keyword{datasets}
