% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{impute_missing_logr}
\alias{impute_missing_logr}
\title{Impute Missing Values}
\usage{
impute_missing_logr(targets, lr_col)
}
\arguments{
\item{targets}{data.table}

\item{lr_col}{Column name with log ratios}
}
\value{
Modified targets with imputed values
}
\description{
Replace NA in log ratio with random noise near 0
}
\keyword{internal}
