% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\name{jumble}
\alias{jumble}
\title{A discrete colour palette}
\usage{
jumble
}
\value{
A character vector.
}
\description{
A pretty discrete palette with 7 colours that are accessible to normal eyes.
The first 5 colours of these are colour-blind safe. Only the first 3 colours
of these are greyscale safe.
}
\examples{
scales::show_col(jumble::jumble)
}
