% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.R
\name{load_bsgenome}
\alias{load_bsgenome}
\title{Load BSgenome Object by Genome Version}
\usage{
load_bsgenome(genome = "hg19")
}
\arguments{
\item{genome}{Character string: "hg19" or "hg38".}
}
\value{
A BSgenome object, or NULL if the package is not available.
}
\description{
Loads the appropriate BSgenome object for the specified genome version.
}
\keyword{internal}
