% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{load_count_files}
\alias{load_count_files}
\title{Load Count Files}
\usage{
load_count_files(count_files)
}
\value{
A list containing count objects, source BED labels, and detected genome builds.
}
\description{
Load Count Files
}
\keyword{internal}
