% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{load_or_generate_counts}
\alias{load_or_generate_counts}
\title{Load or Generate Counts}
\usage{
load_or_generate_counts(bam_file, reference, output_dir)
}
\arguments{
\item{bam_file}{Path to BAM file or counts RDS}

\item{reference}{Reference object}

\item{output_dir}{Output directory}
}
\value{
Counts list
}
\description{
Load or Generate Counts
}
\keyword{internal}
