% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{load_reference_data}
\alias{load_reference_data}
\title{Load Reference Data}
\usage{
load_reference_data(reference_file, annotation_mode = "auto")
}
\arguments{
\item{reference_file}{Path to reference RDS file or reference object}

\item{annotation_mode}{Annotation handling mode.}
}
\value{
Reference object
}
\description{
Load Reference Data
}
\keyword{internal}
