% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{load_target_bed}
\alias{load_target_bed}
\title{Load and Standardize BED File}
\usage{
load_target_bed(target_bed)
}
\value{
A data.table containing standardized target BED records.
}
\description{
Load and Standardize BED File
}
\keyword{internal}
