% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{map_snps_to_bins}
\alias{map_snps_to_bins}
\title{Map SNPs to Target Bins}
\usage{
map_snps_to_bins(snp_table, targets)
}
\arguments{
\item{snp_table}{A data.table of processed SNPs.}

\item{targets}{A data.table of target regions containing 'chromosome', 'start', 'end', 'bin', and 'is_target'.}
}
\value{
A data.table of SNPs filtered and assigned to target bins.
}
\description{
Finds overlaps between extracted SNPs and valid genomic target bins, 
dropping off-target SNPs.
}
\keyword{internal}
