% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somatic.R
\name{map_variants_to_bins}
\alias{map_variants_to_bins}
\title{Map Somatic Variants to Reference Bins}
\usage{
map_variants_to_bins(somatic, reference)
}
\arguments{
\item{somatic}{A processed data.table of somatic variants.}

\item{reference}{The Jumble reference object containing `$ranges`.}
}
\value{
The data.table with an appended `bin` column.
}
\description{
Map Somatic Variants to Reference Bins
}
\keyword{internal}
