% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{match_vcf_sample}
\alias{match_vcf_sample}
\title{Match and Subset VCF by Sample Name}
\usage{
match_vcf_sample(vcf, sample_name = NULL)
}
\arguments{
\item{vcf}{A VCF object.}

\item{sample_name}{Optional string representing the sample name to find.}
}
\value{
A VCF object subsetted to a single sample column.
}
\description{
Determines the correct sample column in a multi-sample VCF based on an exact,
fuzzy, or reverse-fuzzy match of the provided sample name.
}
\keyword{internal}
