% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdis_rman.R
\name{mdis_chrt}
\alias{mdis_chrt}
\title{mdis_chrt: Matrix algebra in base R for covariate balance using Mahalanobis distance of a point to its distribution}
\usage{
mdis_chrt(x_i)
}
\arguments{
\item{x_i}{A vector or covariate matrix}
}
\value{
a numeric vector of length equal to x_i
}
\description{
This function will compute M-distance for a covariate matrix to its mean
}
\examples{
require(jumble)
df_x <- ACTG175[, c('age', 'gender', 'race', 'msm')]

mdis_chrt(df_x)

}
