% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdis_rman.R
\name{mdis_grps}
\alias{mdis_grps}
\title{mdis_grps: Matrix algebra in base R for covariate balance using Mahalanobis distance of two datasets}
\usage{
mdis_grps(x_t, x_c)
}
\arguments{
\item{x_t}{a dataframe or matrix class object, all numeric}

\item{x_c}{a dataframe or matrix class object, all numeric}
}
\value{
a numeric vector of length one, the M-distance multiplied by a constance for sample size
}
\description{
This function will compute M-distance for two datasets by covariate means
}
\examples{
require(jumble)
df_t <- jumble::ACTG175[ACTG175$arms==1, c('age', 'gender', 'race', 'msm')]
df_c <- jumble::ACTG175[ACTG175$arms==0, c('age', 'gender', 'race', 'msm')]

mdis_grps(df_t, df_c)
}
