% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdis_rbase.R
\name{mdis_rbase}
\alias{mdis_rbase}
\title{mdis_rbase: Matrix algebra in base R for covariate balance using Mahalanobis distance of two datasets}
\usage{
mdis_rbase(x_t, x_c)
}
\arguments{
\item{x_t}{a dataframe or matrix class object, all numeric}

\item{x_c}{a dataframe or matrix class object, all numeric}
}
\value{
a numeric vector of length one, the M-distance multiplied by a constance for sample size
}
\description{
This function will compute M-distance for two datasets by covariate means
}
