% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdist.R
\name{mdist}
\alias{mdist}
\title{mdist: Fast calculation for covariate balance using Mahalanobis distance of two datasets}
\usage{
mdist(x_t, x_c)
}
\arguments{
\item{x_t}{a dataframe or matrix class object, all numeric}

\item{x_c}{a dataframe or matrix class object, all numeric}
}
\value{
a numeric vector of length one, the M-distance multiplied by a constance for sample size
}
\description{
This function will compute M-distance for two datasets by covariate means
}
\examples{
require(rerandR)
df_t <- rerandR::ATCG175[ATCG175$arms==1, c('age', 'gender', 'race', 'msm')]
df_c <- rerandR::ATCG175[ATCG175$arms==0, c('age', 'gender', 'race', 'msm')]

mdist(df_t, df_c)

}
