% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{median_correct_backbone}
\alias{median_correct_backbone}
\title{Median Correct by Backbone}
\usage{
median_correct_backbone(targets, lr_col)
}
\arguments{
\item{targets}{data.table with backbone_weight and is_target columns}

\item{lr_col}{Column name with log ratio values}
}
\value{
Modified targets
}
\description{
Subtract weighted median of backbone bins from each sample/target group
}
\keyword{internal}
