% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{navy}
\alias{navy}
\title{Navy colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
navy
}
\description{
Navy colour
}
\examples{
scales::show_col(navy)
}
\keyword{datasets}
