% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_sample}
\alias{normalize_sample}
\title{Normalize Sample Data}
\usage{
normalize_sample(targets, reference_pca)
}
\arguments{
\item{targets}{Query sample data (data.table).}

\item{reference_pca}{PCA results from compute_reference_pca.}
}
\value{
Normalized targets.
}
\description{
Normalizes the query sample using reference PCA and GC correction.
}
