% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somatic.R
\name{parse_vep_csq}
\alias{parse_vep_csq}
\title{Parse VEP CSQ Annotations}
\usage{
parse_vep_csq(vcf, keep_mask)
}
\arguments{
\item{vcf}{A VCF object.}

\item{keep_mask}{Logical vector indicating which rows to process.}
}
\value{
A data.table containing the parsed CSQ columns.
}
\description{
Parses the CSQ string from the VCF INFO column for a subset of variants.
}
\keyword{internal}
