% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{perform_bin_counts}
\alias{perform_bin_counts}
\title{Perform BAM Counting}
\usage{
perform_bin_counts(bam_file, ranges)
}
\value{
A list containing total, short-fragment, medium-fragment, and flag-count vectors.
}
\description{
Perform BAM Counting
}
\keyword{internal}
