% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{perform_segmentation}
\alias{perform_segmentation}
\title{Perform Segmentation}
\usage{
perform_segmentation(targets, reference, alltargets)
}
\arguments{
\item{targets}{Targets data.table}

\item{reference}{Reference object}

\item{alltargets}{Original targets before filtering}
}
\value{
List with targets and segments
}
\description{
Perform Segmentation
}
\keyword{internal}
