% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{piecewise_log2_trans}
\alias{piecewise_log2_trans}
\title{Piecewise Log2 Linear Transform Scale}
\usage{
piecewise_log2_trans()
}
\value{
continuous transformation object
}
\description{
Provides a perfectly joined continuous scale map passing linear `x` below 1 multiplied dynamically scaling visually seamlessly into `log2(x) + offset` bridging perfectly mapping exact distances tracking proportionally visually uniformly visually.
}
\keyword{internal}
