% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{pink}
\alias{pink}
\title{Pink colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
pink
}
\description{
Pink colour
}
\examples{
scales::show_col(pink)
}
\keyword{datasets}
