% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gc_log2}
\alias{plot_gc_log2}
\title{Create GC vs Log2 Plot}
\usage{
plot_gc_log2(targets, yaxis_params, theme_params)
}
\arguments{
\item{targets}{data.table}

\item{yaxis_params}{List from prepare_yaxis_params()}

\item{theme_params}{List from setup_plot_theme()}
}
\value{
ggplot object
}
\description{
Create GC vs Log2 Plot
}
\keyword{internal}
