% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gc_rawdepth}
\alias{plot_gc_rawdepth}
\title{Create GC vs Raw Depth Plot}
\usage{
plot_gc_rawdepth(targets, theme_params, axis_params)
}
\arguments{
\item{targets}{data.table}

\item{theme_params}{List from setup_plot_theme()}

\item{axis_params}{List from prepare_depth_axis()}
}
\value{
ggplot object
}
\description{
Create GC vs Raw Depth Plot
}
\keyword{internal}
