% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_order_log2}
\alias{plot_order_log2}
\title{Create Order vs Log2 Plot}
\usage{
plot_order_log2(
  targets,
  segments,
  cancergenes_summary,
  chroms_order,
  yaxis_params,
  theme_params
)
}
\arguments{
\item{targets}{data.table}

\item{segments}{data.table}

\item{cancergenes_summary}{data.table}

\item{chroms_order}{data.table}

\item{yaxis_params}{List from prepare_yaxis_params()}

\item{theme_params}{List from setup_plot_theme()}
}
\value{
ggplot object
}
\description{
Create Order vs Log2 Plot
}
\keyword{internal}
