% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_order_rawdepth}
\alias{plot_order_rawdepth}
\title{Create Order vs Raw Depth Plot}
\usage{
plot_order_rawdepth(
  targets,
  cancergenes_summary,
  chroms_order,
  theme_params,
  axis_params
)
}
\arguments{
\item{targets}{data.table}

\item{cancergenes_summary}{data.table}

\item{chroms_order}{data.table}

\item{theme_params}{List from setup_plot_theme()}

\item{axis_params}{List from prepare_depth_axis()}
}
\value{
ggplot object
}
\description{
Create Order vs Raw Depth Plot
}
\keyword{internal}
