% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pos_rawdepth}
\alias{plot_pos_rawdepth}
\title{Create Position vs Raw Depth Plot}
\usage{
plot_pos_rawdepth(
  targets,
  cancergenes_summary,
  chroms,
  theme_params,
  axis_params
)
}
\arguments{
\item{targets}{data.table}

\item{cancergenes_summary}{data.table}

\item{chroms}{data.table}

\item{theme_params}{List from setup_plot_theme()}

\item{axis_params}{List from prepare_depth_axis()}
}
\value{
ggplot object
}
\description{
Create Position vs Raw Depth Plot
}
\keyword{internal}
