% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tmb_vaf}
\alias{plot_tmb_vaf}
\title{Plot TMB Estimate vs Minimum VAF Threshold}
\usage{
plot_tmb_vaf(somatic, targets, output_file, title = "TMB VAF Analysis")
}
\arguments{
\item{somatic}{data.table with AF and variants mapping to valid bins.}

\item{targets}{targets data.table.}

\item{output_file}{Path to save the plot (PNG).}

\item{title}{Plot title.}
}
\value{
NULL (saves file).
}
\description{
Plot TMB Estimate vs Minimum VAF Threshold
}
\keyword{internal}
