% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{prepare_bin_metadata}
\alias{prepare_bin_metadata}
\title{Prepare Bin Metadata for Correction}
\usage{
prepare_bin_metadata(targets)
}
\arguments{
\item{targets}{data.table with refmedian columns}
}
\value{
data.table with unique bin-level metadata
}
\description{
Extract reference median values for later normalization
}
\keyword{internal}
