% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_cancergenes_summary}
\alias{prepare_cancergenes_summary}
\title{Prepare Cancer Gene Summary for Labeling}
\usage{
prepare_cancergenes_summary(targets)
}
\arguments{
\item{targets}{data.table with selected_genes}
}
\value{
data.table with summarized cancer genes
}
\description{
Create one point per gene for plot annotations
}
\keyword{internal}
