% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_depth_axis}
\alias{prepare_depth_axis}
\title{Prepare Depth (X-axis) Parameters}
\usage{
prepare_depth_axis()
}
\value{
List with limits, limits_labels, limits_breaks
}
\description{
Setup log-scale depth axis parameters
}
\keyword{internal}
