% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_snp_data}
\alias{prepare_snp_data}
\title{Prepare SNP Data for Plotting}
\usage{
prepare_snp_data(snp_table, targets)
}
\arguments{
\item{snp_table}{data.table with SNP data}

\item{targets}{data.table to map SNPs to}
}
\value{
Modified targets with allele_ratio, DP, and maf columns
}
\description{
Filter and map SNP data to targets
}
\keyword{internal}
