% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_somatic_data}
\alias{prepare_somatic_data}
\title{Prepare Somatic Data for Plotting}
\usage{
prepare_somatic_data(somatic_table, targets, reference = NULL)
}
\arguments{
\item{somatic_table}{data.table with somatic variants}

\item{targets}{data.table to map to}

\item{reference}{Optional reference object with cancergenes}
}
\value{
List with somatic_plot and somatic_labels data.tables
}
\description{
Map somatic variants to targets and create labels
}
\keyword{internal}
