% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_yaxis_params}
\alias{prepare_yaxis_params}
\title{Prepare Y-Axis Parameters}
\usage{
prepare_yaxis_params(targets, segments = NULL)
}
\arguments{
\item{targets}{data.table with smooth_log2 column}
}
\value{
List with ylims, ybreaks, yminorbreaks
}
\description{
Calculate y-axis limits and breaks for log2 plots
}
\keyword{internal}
