% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{process_cancer_genes}
\alias{process_cancer_genes}
\title{Process Local Cancer Genes}
\usage{
process_cancer_genes(allgenes)
}
\value{
A data.table containing display-gene metadata joined to gene coordinates.
}
\description{
Process Local Cancer Genes
}
\keyword{internal}
