% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{process_segments}
\alias{process_segments}
\title{Process and Map Segments to Targets}
\usage{
process_segments(segments, targets)
}
\arguments{
\item{segments}{Segmentation data.table from CBS.}

\item{targets}{Sorted targets data.table.}
}
\value{
A list containing the updated `segments` and `targets` data.tables.
}
\description{
Filters out invalid segments, assigns segment IDs to the target bins, 
and recalculates the median log-ratio for each segment.
}
\keyword{internal}
