% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_strata.R
\name{rand_strata}
\alias{rand_strata}
\title{rnd_strata: Random assignment by strata}
\usage{
rnd_strata(df, strata, seed)
}
\arguments{
\item{df}{a dataframe, containing strata variable}

\item{strata}{a column in df naming the stratification variable, can be character or numeric}

\item{seed}{A seed for random assignment reproducibility}
}
\value{
dataframe with new column assigning treatment
}
\description{
Take a dataset and assign treatment by a specified stratum
}
\examples{

}
