% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{reorder_counts_to_reference}
\alias{reorder_counts_to_reference}
\title{Reorder Counts to Match Reference}
\usage{
reorder_counts_to_reference(counts, reference)
}
\arguments{
\item{counts}{Counts object}

\item{reference}{Reference object}
}
\value{
Reordered counts
}
\description{
Reorder Counts to Match Reference
}
\keyword{internal}
