% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd_allot.R
\name{rnd_allot}
\alias{rnd_allot}
\title{rnd_allot: Random allotment into one of two groups of equal sample size}
\usage{
rnd_allot(x)
}
\arguments{
\item{x}{a Vector to assign}
}
\value{
An array with x, k assignment groups
}
\description{
Used when equal sized groups are desired, if the cohort size is odd then
it will randomly assign the odd one out to one of the groups. Currently only 2
groups.
}
\examples{
require(jumble)
x <- 1:100
rnd_allot(x)

}
