% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd_assign.R
\name{rnd_assign}
\alias{rnd_assign}
\title{rnd_assign: Randomly assign one unit into one of K groups}
\usage{
rnd_assign(x, k, seed)
}
\arguments{
\item{x}{a Vector to assign}

\item{k}{Number of groups to assign}

\item{seed}{A seed for random assignment reproducibility}
}
\value{
An array with x, k assignment groups, and the seed as an attribute
}
\description{
A simple randomizer, provides equal probability of assignment to
one of many groups
}
\examples{
require(jumble)
x <- 1:100
rnd_assign(x, 2, as.integer(as.Date('2016-01-01')))

}
