% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd_rerand.R
\name{rnd_rerand}
\alias{rnd_rerand}
\title{rnd_rerand: Re-randomization on mahalanobis and acceptance criteria}
\usage{
rnd_rerand(x, cutoff, pr_a, seed)
}
\arguments{
\item{x}{an id, vector to assign groups to}

\item{cutoff}{Mahalanobis distance cut-off}

\item{pr_a}{Probability of acceptance, used if cutoff is null; default is 0.001}

\item{seed}{A seed for random assignment reproducibility}
}
\value{
A dataframe, contains id, group assignment, and
attributes giving the seed, cut-off, randomization date, number of iterations
needed to achieve randomization
}
\description{
Conduct a re-randomization procedure using a specified set of covariates,
for each random assignment a mahalanobis distance of group means will be computed
with a sample size adjustment, and the user can either allow the program to find
a cut-off from a chi-square distribution for an given acceptance probability,
default probability is 0.001.  Otherwise specify the cutoff, M-distance.
}
