% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{safe_runmed}
\alias{safe_runmed}
\title{Safe runmed helper}
\usage{
safe_runmed(x, k)
}
\value{
A numeric vector containing running-median smoothed values.
}
\description{
Safe runmed helper
}
\keyword{internal}
