% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{save_analysis_results}
\alias{save_analysis_results}
\title{Save Analysis Results}
\usage{
save_analysis_results(
  targets,
  segments,
  snp_table,
  gis_table,
  qc_metrics,
  output_dir,
  sample_name,
  bam_file,
  reference_file,
  snp_vcf
)
}
\arguments{
\item{targets}{Targets data.table}

\item{segments}{Segments data.table}

\item{snp_table}{SNP table}

\item{gis_table}{GIS table}

\item{qc_metrics}{QC metrics table}

\item{output_dir}{Output directory}

\item{sample_name}{Sample name}

\item{bam_file}{BAM file path}

\item{reference_file}{Reference file path}

\item{snp_vcf}{SNP VCF path}
}
\value{
Invisibly returns NULL after writing result files.
}
\description{
Save Analysis Results
}
\keyword{internal}
