% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{sort_genomically}
\alias{sort_genomically}
\title{Sort Targets Genomically}
\usage{
sort_genomically(targets)
}
\arguments{
\item{targets}{data.table with chromosome and start columns}
}
\value{
Sorted targets
}
\description{
Order by chromosome and genomic position
}
\keyword{internal}
