% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{strip_bottom_axis}
\alias{strip_bottom_axis}
\title{Strip Bottom Axis}
\usage{
strip_bottom_axis(p)
}
\arguments{
\item{p}{ggplot object}
}
\value{
ggplot object with x-axis stripped, but line retained
}
\description{
Strip Bottom Axis
}
\keyword{internal}
