% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{write_qc_metrics}
\alias{write_qc_metrics}
\title{Write QC Metrics to CSV}
\usage{
write_qc_metrics(qc_metrics, output_file)
}
\arguments{
\item{qc_metrics}{QC metrics data.table.}

\item{output_file}{Path to output CSV file.}
}
\value{
Invisibly returns the written QC metrics table.
}
\description{
Writes QC metrics to a CSV file.
}
\keyword{internal}
