% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotate_targets}
\alias{annotate_targets}
\title{Annotate Targets for Plotting}
\usage{
annotate_targets(targets)
}
\arguments{
\item{targets}{data.table.}
}
\value{
Modifies targets in-place.
}
\description{
Adds 'label' and 'selected_genes' columns to targets.
}
\keyword{internal}
