% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{arrange_and_save_plots}
\alias{arrange_and_save_plots}
\title{Arrange and Save Final Plot}
\usage{
arrange_and_save_plots(plot_list, use_snp, title, output_file)
}
\arguments{
\item{plot_list}{List of ggplot objects}

\item{use_snp}{Logical: whether SNP plots are included}

\item{title}{Character: plot title}

\item{output_file}{Character path or NULL}
}
\value{
ggplot object (invisibly if saved)
}
\description{
Arrange and Save Final Plot
}
\keyword{internal}
