% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{build_reference}
\alias{build_reference}
\title{Build Reference Panel}
\usage{
build_reference(
  count_files,
  annotation_source = "biomart",
  genome = NULL,
  output_file = NULL,
  cores = 1,
  mirror = NULL
)
}
\arguments{
\item{count_files}{Vector of paths to count RDS files.}

\item{annotation_source}{Source of annotation ("biomart" or path to folder).}

\item{genome}{Genome version ("hg19" or "hg38"). If NULL, detected from
count files.}

\item{output_file}{Path to save the output RDS file. If NULL, defaults to '<target>.reference.RDS' or 'jumble.WGS.reference.RDS'.}

\item{cores}{Number of cores for parallel processing.}
}
\value{
The reference object (invisibly if saved).
}
\description{
Creates a reference object from a set of count files.
}
