% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_chrom_names}
\alias{clean_chrom_names}
\title{Clean Chromosome Names}
\usage{
clean_chrom_names(x)
}
\arguments{
\item{x}{Vector of chromosome names.}
}
\value{
Cleaned chromosome names (1, 2, ..., X, Y).
}
\description{
Standardizes chromosome names by removing 'chr' prefixes.
}
