% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{comp_gis_for_fraction}
\alias{comp_gis_for_fraction}
\title{Helper to compute GIS for a single fraction}
\usage{
comp_gis_for_fraction(bins_in, tfr)
}
\description{
Helper to compute GIS for a single fraction
}
\keyword{internal}
