% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{compute_gis_and_maf}
\alias{compute_gis_and_maf}
\title{Compute GIS and Add MAF to Targets}
\usage{
compute_gis_and_maf(targets, snp_table, reference)
}
\arguments{
\item{targets}{Targets data.table}

\item{snp_table}{SNP data.table}

\item{reference}{Reference object}
}
\value{
List with gis_table and updated targets
}
\description{
Compute GIS and Add MAF to Targets
}
\keyword{internal}
