% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{correct_by_gc}
\alias{correct_by_gc}
\title{Apply GC Content Correction}
\usage{
correct_by_gc(data, train_indices = NULL, span = 0.75)
}
\arguments{
\item{data}{data.table with 'lr', 'gc' and 'backbone_weight' columns}

\item{train_indices}{Logical or numeric indices for training}

\item{span}{Loess span parameter}
}
\value{
Corrected log ratio vector
}
\description{
Loess smoothing correction for GC content bias
}
\keyword{internal}
