% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{correct_by_optim}
\alias{correct_by_optim}
\title{Apply PCA Correction to Values using Optimisation}
\usage{
correct_by_optim(data, ratio = 1)
}
\arguments{
\item{data}{data.table with 'lr' and PC columns}

\item{ratio}{TV penalty ratio (default 1.0 based on benchmarking)}
}
\value{
Corrected log ratio vector
}
\description{
L1 + Total Variation penalty optimisation using Nelder-Mead
}
\keyword{internal}
