% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{correct_sex_chromosomes}
\alias{correct_sex_chromosomes}
\title{Apply Sex Chromosome Corrections}
\usage{
correct_sex_chromosomes(targets, lr_col)
}
\arguments{
\item{targets}{data.table with male gender assignment}

\item{lr_col}{Column name with log ratio values}
}
\value{
Modified targets with sex-chromosome-corrected values
}
\description{
Adjust X and Y chromosome log ratios for males (nonPAR regions)
}
\keyword{internal}
