% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{define_backbone}
\alias{define_backbone}
\title{Define Backbone Bins}
\usage{
define_backbone(targets, reference = NULL)
}
\arguments{
\item{targets}{Data.table containing bin information.}

\item{reference}{Reference object containing 'mat' and 'bins'.}
}
\value{
Targets with 'is_backbone' column.
}
\description{
Identifies bins to be used as backbone for normalization.
Uses the variance across the reference set to select stable bins.
}
