% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{detect_sample_gender}
\alias{detect_sample_gender}
\title{Detect Sample Gender}
\usage{
detect_sample_gender(targets, lr_col)
}
\arguments{
\item{targets}{data.table}

\item{lr_col}{Column name with log ratio values}
}
\value{
data.table with male, xmedian, ymedian columns
}
\description{
Infer sample gender from X and Y chromosome medians
}
\keyword{internal}
