% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{export_analysis_files}
\alias{export_analysis_files}
\title{Export Analysis Files (CNR, CNS, SEG)}
\usage{
export_analysis_files(targets, segments, output_dir, sample_name)
}
\arguments{
\item{targets}{Targets data.table}

\item{segments}{Segments data.table}

\item{output_dir}{Output directory}

\item{sample_name}{Sample name}
}
\description{
Export Analysis Files (CNR, CNS, SEG)
}
\keyword{internal}
