% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{fetch_ensembl_data}
\alias{fetch_ensembl_data}
\title{Fetch Ensembl Data (Genes or Exons)}
\usage{
fetch_ensembl_data(mart, genome, attributes, type = "genes")
}
\description{
Fetch Ensembl Data (Genes or Exons)
}
\keyword{internal}
