% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_model}
\alias{gis_model}
\title{GIS Model}
\usage{
gis_model(feats = NULL)
}
\arguments{
\item{feats}{List or data.table containing features: focal_gain, local_cnv, loh.}
}
\value{
Predicted GIS score.
}
\description{
Hard-coded prediction function for Genomic Instability Score (GIS).
}
\keyword{internal}
