% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{jumble}
\alias{jumble}
\title{A discrete colour palette}
\usage{
jumble
}
\value{
A character vector.
}
\description{
A pretty discrete 5 colour palette that is relatively colourblind and contrast safe on a light background.
}
\examples{
scales::show_col(jumble)
}
\seealso{
\link{pink} \link{slate}
}
\keyword{datasets}
