% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{jumble}
\alias{jumble}
\title{A discrete colour palette}
\usage{
jumble
}
\value{
A character vector.
}
\description{
A pretty discrete colour palette that is also relatively colourblind and contrast safe on a light background.

Note:
\itemize{
\item The palette is designed for a light background
\item The first 4 colours are colour-blind safe for deutanomaly, protanomaly and tritanomaly
\item The first 5 colours are colour-blind safe for deutanomaly
\item The first 3 colours are desaturated safe.
}
}
\examples{
scales::show_col(jumble)
}
\keyword{datasets}
