% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{jumble}
\alias{jumble}
\title{A discrete colour palette}
\usage{
jumble
}
\value{
A character vector.
}
\description{
A pretty discrete colour palette that is relatively accessible.

Note only the first 3 colours are safe for all forms of colourblindness, as well as greyscale.

An extra safe colour that works is \code{jumble::grey} i.e. \code{jumble[c(1:4, 7)]}.
}
\examples{
scales::show_col(jumble)
}
\keyword{datasets}
