% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{jumble}
\alias{jumble}
\title{A discrete colour palette}
\usage{
jumble
}
\value{
A character vector.
}
\description{
A 7 colour discrete colour palette.

The first 4 colours are colour-blind safe for deutanomaly, protanomaly and tritanomaly.

The first 5 colours are colour-blind safe for deutanomaly.

The first 3 colours are desaturated safe.
}
\examples{
scales::show_col(jumble)
}
\keyword{datasets}
