% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{jumble}
\alias{jumble}
\title{A jumble of useful colours}
\usage{
jumble
}
\value{
A character vector.
}
\description{
A jumble of useful colours.
}
\examples{
scales::show_col(jumble)
}
\keyword{datasets}
