% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{load_reference_data}
\alias{load_reference_data}
\title{Load Reference Data}
\usage{
load_reference_data(reference_file)
}
\arguments{
\item{reference_file}{Path to reference RDS file or reference object}
}
\value{
Reference object
}
\description{
Load Reference Data
}
\keyword{internal}
