% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{prepare_targets}
\alias{prepare_targets}
\title{Prepare Targets Data Table}
\usage{
prepare_targets(reference, counts, bam_file)
}
\arguments{
\item{reference}{Reference object}

\item{counts}{Counts object}

\item{bam_file}{BAM file path}
}
\value{
Targets data.table
}
\description{
Prepare Targets Data Table
}
\keyword{internal}
