% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd_str.R
\name{rnd_str}
\alias{rnd_str}
\title{rnd_str: Random assignment by strata}
\usage{
rnd_str(df, strata)
}
\arguments{
\item{df}{a dataframe, containing strata variable}

\item{strata}{a column in df naming the stratification variable, can be character or numeric}
}
\value{
dataframe with new column assigning treatment
}
\description{
Take a dataset and assign treatment by a specified stratum
}
\examples{

}
