% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{run_cbs}
\alias{run_cbs}
\title{Run Circular Binary Segmentation (CBS)}
\usage{
run_cbs(targets, alpha)
}
\arguments{
\item{targets}{Normalized targets data.table.}

\item{alpha}{Significance level for segmentation.}
}
\value{
A list containing the `segments` data.table and the sorted `targets` data.table.
}
\description{
Prepares the target data by standardizing chromosome names, sorting, 
and running the PSCBS segmentation algorithm.
}
\keyword{internal}
