% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumble.R
\docType{data}
\name{slate}
\alias{slate}
\title{Slate colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
slate
}
\description{
Slate colour
}
\examples{
scales::show_col(slate)
}
\keyword{datasets}
