% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{split_ranges_into_bins}
\alias{split_ranges_into_bins}
\title{Generic Range Splitter
Splits ranges in a data.table into smaller fixed-size bins}
\usage{
split_ranges_into_bins(dt_ranges, bin_size)
}
\arguments{
\item{dt_ranges}{data.table with chromosome, start, end}

\item{bin_size}{integer size of bins}
}
\description{
Generic Range Splitter
Splits ranges in a data.table into smaller fixed-size bins
}
\keyword{internal}
